+++ server/migrations/000006_invite_tokens_ip_rules.down.sql
@@ -1,0 +1,3 @@
+-- Remove invite_tokens and ip_rules tables
+DROP TABLE IF EXISTS ip_rules CASCADE;
+DROP TABLE IF EXISTS invite_tokens CASCADE;